Fully functional Scanner module
Scanner Entry models now have a relationship to Territory models via `territory_id` column in their migration. This means the EntryParseService had to be modified to find the Territory ID when reading the XML scan, and now the `location` column should only contain the coordinates for in-system location, atmosphere and ground (if present). Additionally, Scanne views are now functional and have appropriate routes for all actions.