Improved Scanner module
Created a Job to clean entry logs older than 1 year. Modified the PatrolReminder service to improve the way the Discord message is sent; it now sends relevant information in an inline embed, and attaches a plain text file with the coordinates. Lastly, the Package config.php file has been modified to include the discord webhooks array to use in PatrolReminder and future Notifications.