Add unread reply count badge to discussions
Improved the way that forum_unread_discussions table keeps count of unread replies per user, so it can be displayed in Blade views. Controllers, Models and Traits were tweaked to make use of this feature. Additionally, the Blade views now use the Str facade to limit the names of category/board/discussion in the breadcrumbs to 15 characters; this looks better in mobile versions of the app.