Refactor column names to updated_by and created_by
Models that have an 'author' and/or 'editor' relationship, have the respective columns renamed to 'created_by' and 'updated_by' respectively, for naming convention as well as for using a Trait that updates these fields on Model actions instead of through form validation. This means Form Requests have also been refactored into a single class in most instances.